Add files via upload
a description of Anscombe's quartet and importing matplotlib in order to plot graphs. Some experimentation with plotting the first dataset of the quartet. plotted the graph as 'plt.plot(x, y, 'k')' which just connected each point together and did not provide a linear description. I assigned x to equal all x values of the 1st dataset and y to equal all y values as each dateset only consists of 11 points to plot.