Add files via upload
tried to find the best fit line for the first dataset, with plotting the best fit line, Error message"TypeError: 'numpy.float64' object cannot be interpreted as an integer" was produced. The dataset points are plotting correctly however the best fit line wont work, Used the equation of the line to find the best fit lines as demonstrated in "best fit line video" in the lectures,  I used the code demonstrated in lectures to find the mean of x and y values, the m and c values.